requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-multipart
cachetools
mysqlclient
asyncmy
passlib[bcrypt]
python-dotenv
